+++ strax/plugins/__init__.py
@@ -1,0 +1,7 @@
+from .plugin import *
+from .cut_plugin import *
+from .loop_plugin import *
+from .merge_only_plugin import *
+from .overlap_window_plugin import  *
+from .parrallel_source_plugin import *
+from .rechunk_plugin import *
